LPS-54647 Move out
